+++ FRONT/src/pages/RecipeDetail/RecipeDetail.jsx
@@ -2,12 +2,9 @@ import React, { useEffect, useState, useReducer } from "react";
 import { useAuth0 } from "@auth0/auth0-react";
 import { useDispatch, useSelector } from "react-redux";
 import { useParams, useNavigate, NavLink } from "react-router-dom";
-import {
-  getRecipeDetail,
-  getIngredients,
-  addToCart,
-  setCart,
-} from "../../Redux/actions";
+import { getRecipeDetail } from "../../Redux/actions/recipes";
+import { getIngredients } from "../../Redux/actions/ingredients";
+import { setCart, addToCart } from "../../Redux/actions/cart";
 import s from "../RecipeDetail/RecipeDetail.module.css";
 import NavBar from "../../components/NavBar/NavBar";
 import IngredientsList from "../../components/IngredientsList/ingredientsList";
@@ -30,7 +27,7 @@ const RecipeDetail = () => {
   let { id } = useParams();
   const { user } = useAuth0();
   let dispatch = useDispatch();
-  let recipe = useSelector((state) => state.recipe.recipeDetail);
+  let recipe = useSelector((state) => state.recipes.recipeDetail);
   const ingredients = useSelector((state) => state.ingredients.ingredients);
   const [list, setList] = useState(); // Traigo datos faltantes de ingredients
   //formato list: [{id, name, price}, {id, name, price}...]
@@ -56,7 +53,10 @@ const RecipeDetail = () => {
       localStorage.setItem("MANGIARE_user", JSON.stringify(new_owner));
     } else {
       let new_owner = user ? user.email : "guest";
-      localStorage.setItem("MANGIARE_cart", JSON.stringify([...LS_cart, ...ingredient]));
+      localStorage.setItem(
+        "MANGIARE_cart",
+        JSON.stringify([...LS_cart, ...ingredient])
+      );
       localStorage.setItem("MANGIARE_user", JSON.stringify(new_owner));
     }
   };
@@ -138,7 +138,16 @@ const RecipeDetail = () => {
             filter: "contrast(90%)",
           }}
         >
-          <Text fontSize='6xl' textAlign="center" fontWeight="bold" color="yellow.800" backgroundColor="white" opacity="0.5">{title}</Text>
+          <Text
+            fontSize="6xl"
+            textAlign="center"
+            fontWeight="bold"
+            color="yellow.800"
+            backgroundColor="white"
+            opacity="0.5"
+          >
+            {title}
+          </Text>
 
           <Box width="40%" height="40%" objectFit={"cover"} borderRadius="10px">
             <img className={s.imageDetail} src={image} alt={title} />
@@ -156,7 +165,11 @@ const RecipeDetail = () => {
               <Tab _selected={{ color: "white", bg: "blue.500" }}>Rating</Tab>
             </TabList>
             <TabPanels>
-              <TabPanel backgroundColor="rgba(255, 255, 255, 0.7)" height={["200px", "300px", "400px"]} overflowY="scroll">
+              <TabPanel
+                backgroundColor="rgba(255, 255, 255, 0.7)"
+                height={["200px", "300px", "400px"]}
+                overflowY="scroll"
+              >
                 {!list ? (
                   <h3>Loading...</h3>
                 ) : (
@@ -172,7 +185,7 @@ const RecipeDetail = () => {
                   />
                 )}
               </TabPanel>
-              <TabPanel backgroundColor="rgba(255, 255, 255, 0.5)" width='50%'>
+              <TabPanel backgroundColor="rgba(255, 255, 255, 0.5)" width="50%">
                 <Box width="100%" color="black">
                   <p>{instructions}</p>
                 </Box>
@@ -201,10 +214,11 @@ const RecipeDetail = () => {
           </Tabs>
         </Box>
         <NavLink className={s.navlinkGoBackButton} to={"/home"}>
-        <Button colorScheme="teal" variant="solid" size="lg">Go Home</Button>
-      </NavLink>
+          <Button colorScheme="teal" variant="solid" size="lg">
+            Go Home
+          </Button>
+        </NavLink>
       </Box>
-    
     </>
   );
 };
